Make dom selectors in selectors.js
